# docs/swagger.yaml
basePath: /
definitions:
  entities.AntiScalpingLevel:
    enum:
    - none
    - standard
    - strict
    type: string
    x-enum-varnames:
    - ScalpingNone
    - ScalpingStandard
    - ScalpingStrict
  entities.Appointment:
    properties:
      anti_scalping_level:
        $ref: '#/definitions/entities.AntiScalpingLevel'
      app_code:
        type: string
      attendees_booked:
        type: integer
      booking_duration:
        description: in minutes
        type: integer
      bookings:
        items:
          $ref: '#/definitions/entities.Booking'
        type: array
      created_at:
        type: string
      deleted_at:
        format: date-time
        type: string
      description:
        description: Additional info for the appointment
        type: string
      end_date:
        type: string
      end_time:
        type: string
      id:
        type: string
      max_attendees:
        description: for group appointments
        type: integer
      owner_id:
        type: string
      start_date:
        type: string
      start_time:
        type: string
      status:
        $ref: '#/definitions/entities.AppointmentStatus'
      title:
        type: string
      type:
        $ref: '#/definitions/entities.AppointmentType'
      updated_at:
        type: string
    type: object
  entities.AppointmentStatus:
    enum:
    - pending
    - ongoing
    - completed
    - canceled
    - expired
    type: string
    x-enum-varnames:
    - AppointmentStatusPending
    - AppointmentStatusOngoing
    - AppointmentStatusCompleted
    - AppointmentStatusCanceled
    - AppointmentStatusExpired
  entities.AppointmentType:
    enum:
    - single
    - group
    - party
    type: string
    x-enum-varnames:
    - Single
    - Group
    - Party
...
  entities.Booking:
    properties:
      app_code:
        type: string
      appointment_id:
        type: string
      attendee_count:
        type: integer
      available:
        type: boolean
      booking_code:
        description: Permanent booking code for all bookings
        type: string
      capacity:
        type: integer
      created_at:
        type: string
      date:
        type: string
      deleted_at:
        format: date-time
        type: string
      description:
        description: Additional info from the booker
        type: string
      email:
        type: string
      end_time:
        type: string
      id:
        type: string
      is_slot:
        type: boolean
      name:
        type: string
      notification_channel:
        type: string
      notification_status:
        type: string
      phone:
        type: string
      seats_booked:
        type: integer
      start_time:
        type: string
      status:
        description: 'Booking status: active, cancelled, etc.'
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  errors.APIErrorResponse:
    properties:
      code:
        type: string
      fields:
        items:
          $ref: '#/definitions/errors.FieldError'
        type: array
      message:
        type: string
      meta: {}
      request_id:
        type: string
      status:
        type: integer
    type: object
  errors.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
      rule:
        type: string
    type: object
  requests.AppointmentRequest:
    properties:
      anti_scalping_level:
        allOf:
        - $ref: '#/definitions/entities.AntiScalpingLevel'
        enum:
        - none
        - standard
        - strict
      booking_duration:
        type: integer
      description:
        type: string
      end_date:
        type: string
      end_time:
        type: string
      max_attendees:
        minimum: 1
        type: integer
      start_date:
        type: string
      start_time:
        type: string
      title:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/entities.AppointmentType'
        enum:
        - single
        - group
        - party
    required:
    - booking_duration
    - end_date
    - end_time
    - start_date
    - start_time
    - title
    - type
    type: object
...
  requests.DeviceTokenRequest:
    properties:
      device_id:
        type: string
    required:
    - device_id
    type: object
  requests.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  requests.RefreshTokenRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
...
  responses.AnalyticsResponse:
    properties:
      appointments_by_type:
        additionalProperties:
          type: integer
        description: Breakdowns
        type: object
      avg_attendees_per_booking:
        type: number
      avg_lead_time_hours:
        description: Timing
        type: number
      bookings_by_status:
        additionalProperties:
          type: integer
        type: object
      bookings_per_day:
        description: Time series
        items:
          $ref: '#/definitions/responses.TimeSeriesPoint'
        type: array
      cancellations_per_day:
        items:
          $ref: '#/definitions/responses.TimeSeriesPoint'
        type: array
      distinct_customers:
        type: integer
      end_date:
        type: string
      guest_vs_registered:
        additionalProperties:
          type: integer
        type: object
      median_lead_time_hours:
        type: number
      party_capacity:
        properties:
          percent:
            type: number
          total:
            type: integer
          used:
            type: integer
        type: object
      peak_days:
        items:
          $ref: '#/definitions/responses.BucketCount'
        type: array
      peak_hours:
        description: Insights
        items:
          $ref: '#/definitions/responses.BucketCount'
        type: array
      rejections_per_day:
        items:
          $ref: '#/definitions/responses.TimeSeriesPoint'
        type: array
      repeat_customers:
        type: integer
      slot_utilization_percent:
        description: Utilization & Capacity
        type: number
      start_date:
        type: string
      top_appointments:
        items:
          $ref: '#/definitions/responses.TopAppointment'
        type: array
      total_appointments:
        description: Summary
        type: integer
      total_bookings:
        type: integer
    type: object
  responses.AppointmentResponse:
    properties:
      app_code:
        type: string
      booking_duration:
        type: integer
      created_at:
        type: string
      description:
        type: string
      end_date:
        type: string
      end_time:
        type: string
      id:
        type: string
      max_attendees:
        type: integer
      start_date:
        type: string
      start_time:
        type: string
      status:
        $ref: '#/definitions/entities.AppointmentStatus'
      title:
        type: string
      type:
        $ref: '#/definitions/entities.AppointmentType'
      updated_at:
        type: string
    type: object
  responses.BucketCount:
    properties:
      count:
        type: integer
      key:
        type: string
    type: object
...
  responses.LoginResponse:
    properties:
      expires_in:
        type: integer
      refresh_token:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/responses.UserResponse'
    type: object
  responses.PaginatedResponse:
    properties:
      items:
        items: {}
        type: array
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  responses.TimeSeriesPoint:
    properties:
      count:
        type: integer
      date:
        type: string
    type: object
  responses.TokenResponse:
    properties:
      expires_in:
        type: integer
      refresh_token:
        type: string
      token:
        type: string
    type: object
  responses.TopAppointment:
    properties:
      app_code:
        type: string
      bookings:
        type: integer
      capacity_usage_percent:
        type: number
      title:
        type: string
    type: object
  responses.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a comprehensive API for creating and managing appointments.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Asiko API
  version: "1.0"
paths:
  /analytics:
    get:
      description: |-
        Get analytics for the authenticated user over a date window.
        Includes totals, breakdowns (by type/status, guest vs registered), utilization,
        lead-time stats, daily series, peak hours/days, and top appointments.
      operationId: getUserAnalytics
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.AnalyticsResponse'
        "400":
          description: Invalid date format or missing parameters
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user analytics
      tags:
      - Analytics
...
  /appointments:
    post:
      consumes:
      - application/json
      description: Create a new appointment. Type can be single, group, or party.
      operationId: createAppointment
      parameters:
      - description: Appointment Details
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/requests.AppointmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Appointment'
        "400":
          description: Invalid request payload or validation error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "500":
          description: Failed to create appointment
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new appointment
      tags:
      - Appointments
...
  /appointments/code/{app_code}:
    get:
      description: Retrieves appointment details by its unique app_code, public endpoint
        for booking flow
      operationId: getAppointmentByAppCode
      parameters:
      - description: Appointment identifier (app_code)
        in: path
        name: app_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Appointment'
        "400":
          description: Missing app_code parameter
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "404":
          description: Appointment not found
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      summary: Get appointment by app code
      tags:
      - Appointments
  /appointments/my:
    get:
      description: Retrieves a paginated list of appointments created by the currently
        authenticated user.
      operationId: getMyAppointments
      parameters:
      - collectionFormat: csv
        description: Filter by appointment status (pending, ongoing, completed, canceled,
          expired)
        in: query
        items:
          type: string
        name: status
        type: array
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10)'
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.PaginatedResponse'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/responses.AppointmentResponse'
                  type: array
              type: object
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get appointments created by the user
      tags:
      - Appointments
...
  /auth/device-token:
    post:
      consumes:
      - application/json
      description: Generate a short-lived token for a given device ID to be used in
        booking requests.
      operationId: generateDeviceToken
      parameters:
      - description: Device ID
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/requests.DeviceTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "500":
          description: Could not generate token
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      summary: Generate Device Token
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a refresh token for a new access token
      operationId: refreshToken
      parameters:
      - description: Refresh token
        in: body
        name: refresh
        required: true
        schema:
          $ref: '#/definitions/requests.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.TokenResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "500":
          description: Could not generate token
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      summary: Refresh access token
      tags:
      - Authentication
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and receive a JWT token.
      operationId: loginUser
      parameters:
      - description: Login Credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/requests.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.LoginResponse'
        "202":
          description: Registration pending verification
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "500":
          description: Could not generate token
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      summary: User Login
      tags:
      - Authentication
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer' followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"